Updated the pregame plugin
You can specify one of two modes. Either the old mode by setting
RequireComs = 0. This gives the pregame a fixed time and then starts the
game, regardless of commanders.

The new mode, RequireComs = 1, provides a half-way between the normal
NS2 wait for 2 commanders and the fixed pregame length. If the pregame
extends past your set time and one team has a commander, the game will
start. It will not start if there are no commanders and it will start
normally with 2 before the time's up.