Fixed abrupt start to game
If a 2 commander countdown timer was cancelled whilst the 1 commander is
going, the 1 commander could have reached its end whilst the 2 commander
was going. This would cause the game to immediately start with no
warning.

Fixed by not starting the 2 commander countdown if the 1 commander
countdown has less time left than the 2 commander countdown would last.