Server switcher now has an entry in the vote menu
When a player opens the vote menu, the server will send the current list
of servers to switch to. If a server in the plugin has the flag
"UsersOnly": true, then it will only be sent to registered users. If a
player has no servers sent, the button will never show.