Extendable admin menu
- Plugins can add commands to the commands page in their own category.
Commands can either run the command with a player (or players) argument,
or they can add a popup menu which adds more arguments to the end.
- Plugins can also add new tabs on the side.
- Plugin tabs populate between the standard commands and about tabs in
alphabetical order.
- SGUI controls now support inheriting from other controls.
- Two new SGUI controls, TabPanel which is the base of the admin menu,
and CategoryPanel, which is what is used for the commands list.
- Base commands plugin adds the maps tab, which displays all maps in the
map cycle and allows you to switch to them or call a map vote.
- Bans plugin adds a ban management tab which can add or remove bans.
- Base commands and fun commands add commands to the menu.

This will be documented on the wiki soon for interested plugin authors
(along with the limitations of GUI in NS2).