Team balancing improvements
Everyone believes the new skill system will be their saviour, so we'll
use it when assigning a player to a team to try and join them on the
team they can help the most.

Also, players can now switch teams if there's a team imbalance even if
block teams is set.