Add easier to use time command arguments
Added new "time" command parameter type. Converts a string value into a
time value for easier time values. Units can be specified so that string
based time values are converted from the default, which is seconds.

For ban commands:
1w -> 1 week
2h5m -> 2 hours and 5 minutes

For map vote's extension/round time commands, you can also use full
words/sentences such as "1 hour and 30 minutes", or even "1 hour -5
minutes" which would be 55 minutes.

If the argument given is just a number, than that value is used as
before, with it still being interpreted as raw minutes.