Make the chatbox usable at larger than 1080p res
Tested up to 4K (DSR is pretty handy).

In doing so, fixed the Label control's width/height getters not
accounting for scale.

Also, the position value is now saved so it always opens where you leave
it, as this is needed for larger resolutions. The default position is
mostly off the screen, so you'll need to move it up the first time.