Overhaul UTF8 handling
- GetUTF8Bytes no longer throws an error on failure, it just returns
nil. It also now returns a table of bytes.
- Added UTF8Char, takes a number representing a hex value of a UTF-8
character, and returns the string character for it.
- Added UTF8Encode, takes a string, and for any invalid UTF-8 byte
sequences, replaces them with a dummy character. Returns a table of all
UTF-8 characters in the string, very useful to parse a string's bytes
only once.
- All other UTF8 functions now make use of UTF8Encode, so falling back
to Lua's standard string functions is no longer required.