Fix plugins array having irrelevant plugins
The server now only contains those that can load on the server, and the
client contains only those that can load on the client.

This will save a small bit of processing in hooks, and also allow the
base config to check only server-side extensions are in the active
extensions set.