Improve performance of plugin hook calling
Instead of always going through the whole plugin list and constantly
checking for them being enabled and having a method, build a list of
plugins that have an event on the first call to the event, then use that
thereafter.

If a plugin is loaded or unloaded, simply flush the cache and each
subsequent event call will rebuild it again when needed.