Refactor badges plugin, add tests and master badge lookup
Cleaned up the badges plugin, and added the ability to define a master
badges table.

This table defines which rows certain badges are assigned to globally,
so that group and user entries need only provide a single array and
their badges will be mapped to the row in the master table.