Add external API handler
Handles most of the boilerplate code involved in dealing with external
APIs, including caching (which is temporarily persisted to disk based on
last access time on the server).

APIs and their endpoints are registered using a simple table structure,
which defines how they should be constructed and how their results
should be passed around. Caching requires a cache key to be generated
from the request parameters.