Improve last map exclusion logic
It is now possible to set a minimum and maximum number of maps to
exclude from those recently last played.

When no maximum is provided, as many are removed as possible (in
addition to the minimum) up to the point where the number of maps would
go below the desired options.