Make all-talk state clearer
When global all-talk is enabled, the text should persist on screen
rather than misleadingly stating all-talk is disabled when entering
the PreGame state.

This also makes it interact better with the inventory at the bottom
of the screen, hiding it when visible, or moving the text down if it's
permanently visible.