Add notifications to the admin menu
This makes interacting with the menu provide more clear feedback when
an action has succeeded or not.

The notification system is generic, so may be used by other plugins
too.

This commit also adds the Ionicons font, which provides many useful
icons that may be used in notifications or other UI elements.