Allow eject vote percentage to be based on time commanding
When given a sequence of maximum times and associated percentages,
ejecting a commander becomes based on the total time they have spent
in the command chair for the current round.

This involves overriding the vanilla team votes to avoid them checking
the number of votes needed every tick.

If the sequence contains only one value, the old behaviour is
preserved.