Refactor extension loading mechanism to remove boilerplate
Root files (shared.lua, or server/client.lua if no shared file exists,
or a plugin's single file) now receive the plugin name as an argument.
This can be fed into the new Shine.Plugin() constructor to create a
plugin instance with the appropriate name set.

Non-root files now receive the plugin table and name as arguments,
avoiding the need to access the global 'Plugin' value.

Combining the above, the root file of a plugin can now construct
itself with the appropriate name, and simply return the plugin instance
which will be registered automatically. No more needing to remember to
call Shine:RegisterExtension() or hardcode the plugin name.

This change is entirely backwards compatible. Plugins that rely on
the global 'Plugin' value and manual registration will still work.

Additional changes:
* NetworkUpdate can now be declared in server.lua or client.lua.
* Errors thrown by files loaded with LoadPluginFile are now catchable.
* Modules can now be added before registering a plugin, providing the
plugin is constructed with Shine.Plugin().