Add basic hint system
Registered hints can be set to be displayed up to a given number of
times, with a given delay between them. Hints can also be disabled
entirely if they are no longer relevant.

This is useful to inform users of new functionality without having to
add awkward config values everywhere.