Add friend grouping to shuffle
This allows players to group up together, which informs the optimiser
that they would prefer to be on the same team. The optimiser then
computes a cost for each friend group during optimisation based on how
many players in the group are on the same team.

This cost is weighted using a configured weighting value (similar to
team preferences) and added to the final cost value. This allows the
optimiser to keep friends together if it does not harm the team
balance.

Friend groups have a maximum size limit that is configurable. Larger
groups are less likely to remain together unless there are an
equivalent number of near-equally skilled players to offset them.

This commit also removes the old team optimisation approach entirely,
all optimisation is now entirely cost driven. Further, multiple
iterations are used at lower player counts to allow more solutions to
be explored which produces better teams.