Prevent a round from starting when a map vote starts
This ensures that any pending round start is cancelled, and the round
will not be able to start unless the map vote fails or the change is
vetoed.

This does not however cover corner cases such as a map vote starting
right as a round starts, as there is the potential for a vote to fail
with insufficient votes, in which case the round should continue.