Improve TabPanel behaviour
Add a collapse/expand button in vertical mode when all tabs have icons.
This allows collapsing the tab buttons to only show their icons,
providing room for many more tabs before scrolling is necessary.

Additionally, hide the tab menu button in horizontal mode when the tabs
are not overflowing the container as it is unnecessary.