Avoid opening the map vote over other UI menus
If the mouse is visible at the time of a vote starting and its not for
the vote menu, wait until the mouse is hidden before opening the vote
menu automatically.

This avoids the menu opening over the NS2+ stats menu, as well as any
other menu that uses the mouse.