Make chat messages move above the minimap automatically
When switching to spectate or commanding, a minimap is displayed in the
bottom left of the screen.

The default chat position overlaps this, so as long as the user has not
manually moved their chat (in which case it's their job), automatically
move the chat up above it to avoid it overlapping.