Fix RichText scaling issues
Use a single dummy element with 0 size as the root, and place all
elements as children of it. The previous method meant that scaled text
at the start of a line would scale all other elements after it.

The dummy element is used to control alpha over all rich text elements.