Fix invisible labels under a cropping element
The cropping logic does not compute the text's actual size, and thus
if a label starts at a negative x/y relative to a cropping element
and has not yet been rendered it will not render at all.

This forces the text's internal size to be computed whenever anything
that could change it is updated.