Add indication of friend group team preference
When in a friend group, the group shares a team preference based on
what the majority in the group want. If there's an even split between
the 2 teams, the preference is ignored.

This adds a message when the group's preference deviates from the local
player's preference, as well as adding an indicator in the vote menu for
the current group preference.