Add support for per team/commander skill values
Commander skills are used when a player is a commander at the time of
the shuffle, and they are not being swapped to the opposite team.

Per team skills are used both at the commander level and player level
when enabled.

Displaying standard deviation has to be disabled when using these new
values as they are not currently networked.