Add support for prefixed map names
Certain map names are treated differently by the game to load a certain
mod for the map. Previously the map vote system didn't know about this,
and thus such maps had no preview images and were not detected as the
current map.

Now if the mod for a prefix is loaded, the current map is assumed to be
the prefixed variant, and thus the map vote correctly identifies which
map to extend if both the prefixed and non-prefixed maps are available
(as well as the appropriate map settings and position in the cycle).

This also adds the capability to overlay a texture over a map preview
for prefixed maps by adding a texture file to ui/shine/map_overlays.
An image is included for the "infest" prefix (contributed by ailmanki
on GitHub).