Add automatic alpha inheritance with multiplier
When alpha inheritance propagation is enabled and alpha is set to be
inherited, automatically compensate for parent alpha values to ensure
that the alpha requested in a given colour is the alpha that an item
renders with.

This removes the tedium of having to manually compensate for alpha
inheritance, making it all seamlessly automatic.