Remove beta extensions concept
This was an attempt to stage plugins that were new until they were
stable enough to enable by default, but it ends up polluting the config.

Instead, avoid adding new plugins to the config until either an admin
enables/disables them, or the plugin sets a default state. This provides
the same staging functionality without the extra config.